+++ Runtime/Default/UnityEngine/DefaultSystemInfo.cs
@@ -1,0 +1,199 @@
+namespace Moqunity.UnityEngine
+{
+	public partial class DefaultSystemInfo : Abstract.UnityEngine.SystemInfo
+	{
+		public string unsupportedIdentifier => global::UnityEngine.SystemInfo.unsupportedIdentifier;
+
+		public float batteryLevel => global::UnityEngine.SystemInfo.batteryLevel;
+
+		public global::UnityEngine.BatteryStatus batteryStatus => global::UnityEngine.SystemInfo.batteryStatus;
+
+		public string operatingSystem => global::UnityEngine.SystemInfo.operatingSystem;
+
+		public global::UnityEngine.OperatingSystemFamily operatingSystemFamily => global::UnityEngine.SystemInfo.operatingSystemFamily;
+
+		public string processorType => global::UnityEngine.SystemInfo.processorType;
+
+		public int processorFrequency => global::UnityEngine.SystemInfo.processorFrequency;
+
+		public int processorCount => global::UnityEngine.SystemInfo.processorCount;
+
+		public int systemMemorySize => global::UnityEngine.SystemInfo.systemMemorySize;
+
+		public string deviceUniqueIdentifier => global::UnityEngine.SystemInfo.deviceUniqueIdentifier;
+
+		public string deviceName => global::UnityEngine.SystemInfo.deviceName;
+
+		public string deviceModel => global::UnityEngine.SystemInfo.deviceModel;
+
+		public bool supportsAccelerometer => global::UnityEngine.SystemInfo.supportsAccelerometer;
+
+		public bool supportsGyroscope => global::UnityEngine.SystemInfo.supportsGyroscope;
+
+		public bool supportsLocationService => global::UnityEngine.SystemInfo.supportsLocationService;
+
+		public bool supportsVibration => global::UnityEngine.SystemInfo.supportsVibration;
+
+		public bool supportsAudio => global::UnityEngine.SystemInfo.supportsAudio;
+
+		public global::UnityEngine.DeviceType deviceType => global::UnityEngine.SystemInfo.deviceType;
+
+		public int graphicsMemorySize => global::UnityEngine.SystemInfo.graphicsMemorySize;
+
+		public string graphicsDeviceName => global::UnityEngine.SystemInfo.graphicsDeviceName;
+
+		public string graphicsDeviceVendor => global::UnityEngine.SystemInfo.graphicsDeviceVendor;
+
+		public int graphicsDeviceID => global::UnityEngine.SystemInfo.graphicsDeviceID;
+
+		public int graphicsDeviceVendorID => global::UnityEngine.SystemInfo.graphicsDeviceVendorID;
+
+		public global::UnityEngine.Rendering.GraphicsDeviceType graphicsDeviceType => global::UnityEngine.SystemInfo.graphicsDeviceType;
+
+		public bool graphicsUVStartsAtTop => global::UnityEngine.SystemInfo.graphicsUVStartsAtTop;
+
+		public string graphicsDeviceVersion => global::UnityEngine.SystemInfo.graphicsDeviceVersion;
+
+		public int graphicsShaderLevel => global::UnityEngine.SystemInfo.graphicsShaderLevel;
+
+		public bool graphicsMultiThreaded => global::UnityEngine.SystemInfo.graphicsMultiThreaded;
+
+		public global::UnityEngine.Rendering.RenderingThreadingMode renderingThreadingMode => global::UnityEngine.SystemInfo.renderingThreadingMode;
+
+		public bool hasHiddenSurfaceRemovalOnGPU => global::UnityEngine.SystemInfo.hasHiddenSurfaceRemovalOnGPU;
+
+		public bool hasDynamicUniformArrayIndexingInFragmentShaders => global::UnityEngine.SystemInfo.hasDynamicUniformArrayIndexingInFragmentShaders;
+
+		public bool supportsShadows => global::UnityEngine.SystemInfo.supportsShadows;
+
+		public bool supportsRawShadowDepthSampling => global::UnityEngine.SystemInfo.supportsRawShadowDepthSampling;
+
+		public bool supportsMotionVectors => global::UnityEngine.SystemInfo.supportsMotionVectors;
+
+		public bool supports3DTextures => global::UnityEngine.SystemInfo.supports3DTextures;
+
+		public bool supportsCompressed3DTextures => global::UnityEngine.SystemInfo.supportsCompressed3DTextures;
+
+		public bool supports2DArrayTextures => global::UnityEngine.SystemInfo.supports2DArrayTextures;
+
+		public bool supports3DRenderTextures => global::UnityEngine.SystemInfo.supports3DRenderTextures;
+
+		public bool supportsCubemapArrayTextures => global::UnityEngine.SystemInfo.supportsCubemapArrayTextures;
+
+		public global::UnityEngine.Rendering.CopyTextureSupport copyTextureSupport => global::UnityEngine.SystemInfo.copyTextureSupport;
+
+		public bool supportsComputeShaders => global::UnityEngine.SystemInfo.supportsComputeShaders;
+
+		public bool supportsConservativeRaster => global::UnityEngine.SystemInfo.supportsConservativeRaster;
+
+		public bool supportsMultiview => global::UnityEngine.SystemInfo.supportsMultiview;
+
+		public bool supportsGeometryShaders => global::UnityEngine.SystemInfo.supportsGeometryShaders;
+
+		public bool supportsTessellationShaders => global::UnityEngine.SystemInfo.supportsTessellationShaders;
+
+		public bool supportsRenderTargetArrayIndexFromVertexShader => global::UnityEngine.SystemInfo.supportsRenderTargetArrayIndexFromVertexShader;
+
+		public bool supportsInstancing => global::UnityEngine.SystemInfo.supportsInstancing;
+
+		public bool supportsHardwareQuadTopology => global::UnityEngine.SystemInfo.supportsHardwareQuadTopology;
+
+		public bool supports32bitsIndexBuffer => global::UnityEngine.SystemInfo.supports32bitsIndexBuffer;
+
+		public bool supportsSparseTextures => global::UnityEngine.SystemInfo.supportsSparseTextures;
+
+		public int supportedRenderTargetCount => global::UnityEngine.SystemInfo.supportedRenderTargetCount;
+
+		public bool supportsSeparatedRenderTargetsBlend => global::UnityEngine.SystemInfo.supportsSeparatedRenderTargetsBlend;
+
+		public int supportedRandomWriteTargetCount => global::UnityEngine.SystemInfo.supportedRandomWriteTargetCount;
+
+		public int supportsMultisampledTextures => global::UnityEngine.SystemInfo.supportsMultisampledTextures;
+
+		public bool supportsMultisampled2DArrayTextures => global::UnityEngine.SystemInfo.supportsMultisampled2DArrayTextures;
+
+		public bool supportsMultisampleAutoResolve => global::UnityEngine.SystemInfo.supportsMultisampleAutoResolve;
+
+		public int supportsTextureWrapMirrorOnce => global::UnityEngine.SystemInfo.supportsTextureWrapMirrorOnce;
+
+		public bool usesReversedZBuffer => global::UnityEngine.SystemInfo.usesReversedZBuffer;
+
+		public global::UnityEngine.NPOTSupport npotSupport => global::UnityEngine.SystemInfo.npotSupport;
+
+		public int maxTextureSize => global::UnityEngine.SystemInfo.maxTextureSize;
+
+		public int maxCubemapSize => global::UnityEngine.SystemInfo.maxCubemapSize;
+
+		public int maxComputeBufferInputsVertex => global::UnityEngine.SystemInfo.maxComputeBufferInputsVertex;
+
+		public int maxComputeBufferInputsFragment => global::UnityEngine.SystemInfo.maxComputeBufferInputsFragment;
+
+		public int maxComputeBufferInputsGeometry => global::UnityEngine.SystemInfo.maxComputeBufferInputsGeometry;
+
+		public int maxComputeBufferInputsDomain => global::UnityEngine.SystemInfo.maxComputeBufferInputsDomain;
+
+		public int maxComputeBufferInputsHull => global::UnityEngine.SystemInfo.maxComputeBufferInputsHull;
+
+		public int maxComputeBufferInputsCompute => global::UnityEngine.SystemInfo.maxComputeBufferInputsCompute;
+
+		public int maxComputeWorkGroupSize => global::UnityEngine.SystemInfo.maxComputeWorkGroupSize;
+
+		public int maxComputeWorkGroupSizeX => global::UnityEngine.SystemInfo.maxComputeWorkGroupSizeX;
+
+		public int maxComputeWorkGroupSizeY => global::UnityEngine.SystemInfo.maxComputeWorkGroupSizeY;
+
+		public int maxComputeWorkGroupSizeZ => global::UnityEngine.SystemInfo.maxComputeWorkGroupSizeZ;
+
+		public bool supportsAsyncCompute => global::UnityEngine.SystemInfo.supportsAsyncCompute;
+
+		public bool supportsGpuRecorder => global::UnityEngine.SystemInfo.supportsGpuRecorder;
+
+		public bool supportsGraphicsFence => global::UnityEngine.SystemInfo.supportsGraphicsFence;
+
+		public bool supportsAsyncGPUReadback => global::UnityEngine.SystemInfo.supportsAsyncGPUReadback;
+
+		public bool supportsRayTracing => global::UnityEngine.SystemInfo.supportsRayTracing;
+
+		public bool supportsSetConstantBuffer => global::UnityEngine.SystemInfo.supportsSetConstantBuffer;
+
+		public int constantBufferOffsetAlignment => global::UnityEngine.SystemInfo.constantBufferOffsetAlignment;
+
+		public bool minConstantBufferOffsetAlignment => global::UnityEngine.SystemInfo.minConstantBufferOffsetAlignment;
+
+		public bool hasMipMaxLevel => global::UnityEngine.SystemInfo.hasMipMaxLevel;
+
+		public bool supportsMipStreaming => global::UnityEngine.SystemInfo.supportsMipStreaming;
+
+		public bool usesLoadStoreActions => global::UnityEngine.SystemInfo.usesLoadStoreActions;
+
+		public global::UnityEngine.HDRDisplaySupportFlags hdrDisplaySupportFlags => global::UnityEngine.SystemInfo.hdrDisplaySupportFlags;
+
+		public bool supportsStoreAndResolveAction => global::UnityEngine.SystemInfo.supportsStoreAndResolveAction;
+
+		public bool supportsRenderToCubemap => global::UnityEngine.SystemInfo.supportsRenderToCubemap;
+
+		public bool supportsImageEffects => global::UnityEngine.SystemInfo.supportsImageEffects;
+
+		public int supportsStencil => global::UnityEngine.SystemInfo.supportsStencil;
+
+		public int graphicsPixelFillrate => global::UnityEngine.SystemInfo.graphicsPixelFillrate;
+
+		public bool supportsVertexPrograms => global::UnityEngine.SystemInfo.supportsVertexPrograms;
+
+		public global::UnityEngine.Experimental.Rendering.GraphicsFormat GetCompatibleFormat(global::UnityEngine.Experimental.Rendering.GraphicsFormat format, global::UnityEngine.Experimental.Rendering.FormatUsage usage) => global::UnityEngine.SystemInfo.GetCompatibleFormat(format, usage);
+
+		public global::UnityEngine.Experimental.Rendering.GraphicsFormat GetGraphicsFormat(global::UnityEngine.Experimental.Rendering.DefaultFormat format) => global::UnityEngine.SystemInfo.GetGraphicsFormat(format);
+
+		public int GetRenderTextureSupportedMSAASampleCount(global::UnityEngine.RenderTextureDescriptor desc) => global::UnityEngine.SystemInfo.GetRenderTextureSupportedMSAASampleCount(desc);
+
+		public bool IsFormatSupported(global::UnityEngine.Experimental.Rendering.GraphicsFormat format, global::UnityEngine.Experimental.Rendering.FormatUsage usage) => global::UnityEngine.SystemInfo.IsFormatSupported(format, usage);
+
+		public bool SupportsBlendingOnRenderTextureFormat(global::UnityEngine.RenderTextureFormat format) => global::UnityEngine.SystemInfo.SupportsBlendingOnRenderTextureFormat(format);
+
+		public bool SupportsRenderTextureFormat(global::UnityEngine.RenderTextureFormat format) => global::UnityEngine.SystemInfo.SupportsRenderTextureFormat(format);
+
+		public bool SupportsTextureFormat(global::UnityEngine.TextureFormat format) => global::UnityEngine.SystemInfo.SupportsTextureFormat(format);
+
+		public bool SupportsVertexAttributeFormat(global::UnityEngine.Rendering.VertexAttributeFormat format, int dimension) => global::UnityEngine.SystemInfo.SupportsVertexAttributeFormat(format, dimension);
+	}
+}