+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Input interface and implementation
 - Screen interface and implementation
 - Time interface and implementation
+- SystemInfo interface and implementation
 
 ## [0.1.0] - 2021-10-10
 ### Added